# tripmatch_backend/docs/api-contract-iteratia-1-2.openapi.yaml
openapi: 3.0.3
info:
  title: TripMatch Backend API Contract (Iteratia 1 + 2 + 3)
  version: "iter1-iter2-iter3"
  description: >
    Contract extras direct din codebase (controller + DTO + validari service).
    Toate endpointurile protejate folosesc Basic Auth, exceptand /api/auth/**.
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
...
  /api/trips:
    get:
      tags: [Trip]
      summary: List trips for authenticated user
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault10"
        - $ref: "#/components/parameters/sortParamTrips"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageTripListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Trip]
      summary: Create trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
    get:
      tags: [Trip]
      summary: Get trip detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Trip]
      summary: Update trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Trip]
      summary: Delete trip
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}/budget:
    parameters:
      - $ref: "#/components/parameters/positiveTripIdPath"
    get:
      tags: [Budget]
      summary: Get trip budget
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripBudgetResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Budget]
      summary: Upsert trip budget + optional category budgets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTripBudgetRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripBudgetResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}/destinations:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
    get:
      tags: [Destination]
      summary: List destinations in trip
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault10"
        - $ref: "#/components/parameters/sortParamDestinations"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDestinationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Destination]
      summary: Create destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDestinationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}/destinations/{destinationId}:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
      - $ref: "#/components/parameters/destinationIdPath"
    get:
      tags: [Destination]
      summary: Get destination detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Destination]
      summary: Update destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDestinationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Destination]
      summary: Delete destination
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}/destinations/{destinationId}/activities:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
      - $ref: "#/components/parameters/destinationIdPath"
    get:
      tags: [Activity]
      summary: List activities
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault10"
        - $ref: "#/components/parameters/sortParamActivities"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageActivityListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Activity]
      summary: Create activity
      description: >
        Daca allowOverlap != true, backend verifica suprapuneri orare in aceeasi zi;
        la conflict returneaza 409 cu lista id-urilor in mesaj.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/trips/{tripId}/destinations/{destinationId}/activities/reorder:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
      - $ref: "#/components/parameters/destinationIdPath"
    put:
      tags: [Activity]
      summary: Reorder activities (UC-08)
      description: >
        orderedActivityIds trebuie sa contina TOATE activitatile destinatiei, fara duplicate.
        Reordonarea este aplicata per zi (activityDate). Raspunsul are id = null.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderActivitiesRequest"
            example:
              orderedActivityIds: [101, 104, 102, 103]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
              example:
                id: null
                message: "Activities reordered successfully!"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}/destinations/{destinationId}/activities/{activityId}:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
      - $ref: "#/components/parameters/destinationIdPath"
      - $ref: "#/components/parameters/activityIdPath"
    get:
      tags: [Activity]
      summary: Get activity detail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Activity]
      summary: Update activity
      description: >
        Daca allowOverlap != true, backend verifica suprapuneri dupa valorile efective (request + existing).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Activity]
      summary: Delete activity
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/{tripId}/destinations/{destinationId}/activities/{activityId}/duplicate:
    parameters:
      - $ref: "#/components/parameters/tripIdPath"
      - $ref: "#/components/parameters/destinationIdPath"
      - $ref: "#/components/parameters/activityIdPath"
    post:
      tags: [Activity]
      summary: Duplicate activity (UC-08)
      description: >
        request body este optional. Daca lipseste, copiaza in aceeasi data/ora.
        Daca allowOverlap != true, backend verifica suprapuneri.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicateActivityRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/trips/ai/generate:
    post:
      tags: [AI Draft]
      summary: Generate AI draft trip
      description: >
        destinationHint sau tripType trebuie furnizat.
        Erorile OpenAI sunt mapate la 503.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAiDraftRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/trips/ai/drafts:
    get:
      tags: [AI Draft]
      summary: List AI drafts (current user)
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault10"
        - $ref: "#/components/parameters/sortParamAiDrafts"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAiDraftListItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/ai/drafts/{id}:
    parameters:
      - $ref: "#/components/parameters/positiveIdPath"
    get:
      tags: [AI Draft]
      summary: Get draft by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [AI Draft]
      summary: Update AI draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAiDraftRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/ai/drafts/{id}/save-as-trip:
    parameters:
      - $ref: "#/components/parameters/positiveIdPath"
    post:
      tags: [AI Draft]
      summary: Persist draft as real trip
      description: >
        request body este optional.
        Daca tripTitleOverride este trimis, trebuie sa nu fie blank dupa trim.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveDraftAsTripRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/trips/ai/drafts/{id}/regenerate-day:
    parameters:
      - $ref: "#/components/parameters/positiveIdPath"
    post:
      tags: [AI Draft]
      summary: Regenerate one day from draft
      description: Erorile OpenAI sunt mapate la 503.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateDraftDayRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    pageParamDefault0:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    sizeParamDefault10:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 10
...
    sortParamTrips:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: "createdAt,desc"
      description: Spring pageable sort, ex `createdAt,desc`.
    sortParamDestinations:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: "createdAt,desc"
      description: Spring pageable sort, ex `createdAt,desc`.
...
    tripIdPath:
      name: tripId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    positiveTripIdPath:
      name: tripId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    destinationIdPath:
      name: destinationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: >
        Bad Request. In acest backend, body-ul poate fi:
        1) string simplu (Business error) sau
        2) array de stringuri (Bean validation errors).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationOrMessageError"
    Unauthorized:
      description: Missing/invalid Basic Auth credentials.
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: string
          example: "Trip not found!"
...
  schemas:
...
    CreateTripRequest:
      type: object
      description: endDate must be strictly after startDate.
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 255
          pattern: "^.*\\S.*$"
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdWithAi:
          type: boolean
      required: [title, startDate, endDate, createdWithAi]
    UpdateTripRequest:
      type: object
      description: If both dates are present, endDate cannot be before startDate.
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 255
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    TripResponse:
      type: object
      properties:
        tripId: { type: integer, format: int64 }
        message: { type: string }
      required: [tripId, message]
...
    TripDetailResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string, nullable: true }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        createdWithAi: { type: boolean }
        status:
          $ref: "#/components/schemas/TripStatus"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateDestinationRequest:
      type: object
      description: endDate cannot be before startDate.
      properties:
        city:
          type: string
          minLength: 2
          maxLength: 125
          pattern: "^.*\\S.*$"
        country:
          type: string
          minLength: 2
          maxLength: 125
          pattern: "^.*\\S.*$"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        notes:
          type: string
      required: [city, country, startDate, endDate]
...
    DestinationResponse:
      type: object
      properties:
        destinationId: { type: integer, format: int64 }
        message: { type: string }
      required: [destinationId, message]
...
    DestinationDetailResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        city: { type: string }
        country: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UpsertTripBudgetRequest:
      type: object
      description: >
        If categories is present, duplicate category values are rejected (400),
        and existing categories missing from payload are deleted.
      properties:
        totalBudget:
          type: number
          minimum: 0.01
        currency:
          type: string
          maxLength: 64
          pattern: "^.*\\S.*$"
        spentTotal:
          type: number
          minimum: 0
          nullable: true
        categories:
          type: array
          maxItems: 20
          nullable: true
          items:
            $ref: "#/components/schemas/UpsertTripBudgetCategoryRequest"
      required: [totalBudget, currency]
...
    TripBudgetResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tripId: { type: integer, format: int64 }
        totalBudget: { type: number }
        currency: { type: string }
        estimatedTotal: { type: number }
        spentTotal: { type: number }
        remainingByEstimated: { type: number }
        remainingBySpent: { type: number }
        estimatedProgressPct: { type: number }
        spentProgressPct: { type: number }
        categories:
          type: array
          items:
            $ref: "#/components/schemas/TripBudgetCategoryResponse"
...
    PageTripListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TripListResponse"
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
      required: [content, page, size, totalElements, totalPages, last]
    PageDestinationListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/DestinationListResponse"
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
      required: [content, page, size, totalElements, totalPages, last]
...